+++ CHANGELOG.md
@@ -1,3 +1,11 @@
+### 1.0.3 (2025-03-15)
+
+animations on show
+more locations (top, middle, bottom -> left, center, right)
+offset for the button
+offset for the icon inside the button
+disabled state
+
 ### 1.0.2 (2024-08-15)
 
 support for "fire-dom-event" and "url"